Use serde instead of rustc-serialize
Fixes #4
